+++ src/helpful-queries/patch_2024-02-29.sql
@@ -75,3 +75,49 @@ WHERE
           COUNT(DISTINCT BuildItems.itemId) = 1
       ) AS tmp
   );
+
+# Flag builds that contain Twisted Arbalest + Bore
+UPDATE
+  Build
+SET
+  isPatchAffected = true
+WHERE
+  id IN (
+    SELECT
+      id
+    FROM
+      (
+        SELECT
+          Build.id
+        FROM
+          Build
+          JOIN BuildItems ON Build.id = BuildItems.buildId
+        WHERE
+          BuildItems.itemId IN ('13hsq0', 'iq4wjy')
+          AND Build.createdAt < '2024-02-29'
+        GROUP BY
+          Build.id
+        HAVING
+          COUNT(DISTINCT BuildItems.itemId) = 2
+      ) AS tmp
+  );
+
+# Get count of affected builds for a particular updatedAt
+SELECT
+  COUNT(*)
+FROM
+  (
+    SELECT
+      Build.id
+    FROM
+      Build
+      JOIN BuildItems ON Build.id = BuildItems.buildId
+    WHERE
+      BuildItems.itemId IN ('13hsq0', 'iq4wjy')
+      AND Build.createdAt < '2024-02-29'
+      AND Build.isPatchAffected = false
+    GROUP BY
+      Build.id
+    HAVING
+      COUNT(DISTINCT BuildItems.itemId) = 2
+  ) AS tmp;